Refactored large main file into directory
