+++ client-app/src/desktop/tabs/containers/tab/index.js
@@ -1,0 +1,11 @@
+/*
+ * This file belongs to Hoist, an application development toolkit
+ * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
+ *
+ * Copyright © 2021 Extremely Heavy Industries Inc.
+ */
+
+export * from './CustomTabContainerPanel';
+export * from './TabStateTabContainerPanel';
+export * from './DynamicTabContainerPanel';
+export * from './createContainerModelConfig';
